app/backend/api-agent/requirements.txt
# Framework
litestar[standard]
uvicorn

# LangChain & AI
langchain-google-genai
google-genai
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-postgres==0.0.12

# Database
asyncpg
sqlalchemy
psycopg2-binary

# Utils
python-dotenv
pydantic-settings
